+++ Sources/Sentry/include/SentryCrashWrapper.h
@@ -1,5 +1,5 @@
 #import "SentryDefines.h"
-#import "SentryInternalDefines.h"
+#import "SentryInternalCDefines.h"
 #import <Foundation/Foundation.h>
 
 NS_ASSUME_NONNULL_BEGIN