+++ Tests/SentryTests/Protocol/SentryThreadTests.swift
@@ -17,4 +17,12 @@ class SentryThreadTests: XCTestCase {
         XCTAssertNotNil(actual["stacktrace"])
         XCTAssertTrue(actual["main"] as! Bool)
     }
+    
+    func testSerialize_Bools() {
+        let thread = SentryThread(threadId: 0)
+        
+        SentryBooleanSerialization.test(thread, property: "crashed")
+        SentryBooleanSerialization.test(thread, property: "current")
+        SentryBooleanSerialization.test(thread, property: "isMain", serializedProperty: "main")
+    }
 }