+++ Sources/Sentry/SentryProfiler.mm
@@ -16,6 +16,7 @@
 #    import "SentryHexAddressFormatter.h"
 #    import "SentryHub+Private.h"
 #    import "SentryId.h"
+#    import "SentryInternalDefines.h"
 #    import "SentryLog.h"
 #    import "SentryMetricProfiler.h"
 #    import "SentryNSProcessInfoWrapper.h"
@@ -629,14 +630,7 @@ processFrameRates(SentryFrameInfoTimeSeries *frameRates, uint64_t start)
     const auto debugImages = [NSMutableArray<NSDictionary<NSString *, id> *> new];
     const auto debugMeta = [_debugImageProvider getDebugImages];
     for (SentryDebugMeta *debugImage in debugMeta) {
-        const auto debugImageDict = [NSMutableDictionary<NSString *, id> dictionary];
-        debugImageDict[@"type"] = @"macho";
-        debugImageDict[@"debug_id"] = debugImage.uuid;
-        debugImageDict[@"code_file"] = debugImage.name;
-        debugImageDict[@"image_addr"] = debugImage.imageAddress;
-        debugImageDict[@"image_size"] = debugImage.imageSize;
-        debugImageDict[@"image_vmaddr"] = debugImage.imageVmAddress;
-        [debugImages addObject:debugImageDict];
+        [debugImages addObject:[debugImage serialize]];
     }
     if (debugImages.count > 0) {
         profile[@"debug_meta"] = @{ @"images" : debugImages };