+++ CHANGELOG.md
@@ -16,6 +16,11 @@
 - Fix atomic import error for profiling (#2683)
 - Don't create breadcrumb for UITextField editingChanged event (#2686)
 - Fix EXC_BAD_ACCESS in SentryTracer (#2697)
+- Serialization of nullable booleans (#2706)
+
+### Improvements
+
+- Change debug image type to macho (#2701)
 
 ## 8.1.0