+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 ### Features
 
 - Combine UIKit and SwiftUI transactions (#2681)
+- Add isMain thread to SentryThread (#2692)
 - Add enableTracing option (#2693)
+- Add `in_foreground` to App Context (#2692)
 
 ### Fixes
 
@@ -14,6 +16,12 @@
 - Keep status of auto transactions when finishing (#2684)
 - Fix atomic import error for profiling (#2683)
 - Don't create breadcrumb for UITextField editingChanged event (#2686)
+- Fix EXC_BAD_ACCESS in SentryTracer (#2697)
+- Serialization of nullable booleans (#2706)
+
+### Improvements
+
+- Change debug image type to macho (#2701)
 
 ## 8.1.0