+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 8.2.0
 
 - Add cache directory option ([#2711](https://github.com/getsentry/sentry-cocoa/pull/2711))